Revert "add an additional thread to manage server requests"
This reverts commit 0e1d8fdd19587afbe8d82f7f4684ad2900a65663.

+++ include/Query.cpp
@@ -63,7 +63,7 @@ Query::Query(const std::vector<std::string>& tokens) {
       } else if ((*it) == "triggers") {
          type = TRIGGER;
 
-         triggers_info.frequency = std::stoi(string_util::next_token(it));         
+         triggers_info.frequency = std::stof(string_util::next_token(it));         
       }
    }
 }
